# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cakapku API Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3456",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/periodes": {
            "get": {
                "tags": [
                    "Periode"
                ],
                "summary": "Get list of Periode",
                "operationId": "ab2b4217ed75f5e226c1321ff0e444a4",
                "responses": {
                    "200": {
                        "description": "List of Periode"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Periode"
                ],
                "summary": "Create new Periode",
                "operationId": "e10258e55fa81bda17af0a298c38cd1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tahun",
                                    "status"
                                ],
                                "properties": {
                                    "tahun": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Periode created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/periodes/{id}": {
            "get": {
                "tags": [
                    "Periode"
                ],
                "summary": "Get specific Periode",
                "operationId": "e406b3c501e093573847b789de57162d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Periode object"
                    },
                    "404": {
                        "description": "Periode not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Periode"
                ],
                "summary": "Update Periode",
                "operationId": "2dec592ba9fa8b1780192fd7bc66b2df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tahun": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Periode updated successfully"
                    },
                    "404": {
                        "description": "Periode not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Periode"
                ],
                "summary": "Delete Periode",
                "operationId": "b177c9df6a0a065c97667a7953879e7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Periode deleted successfully"
                    },
                    "404": {
                        "description": "Periode not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}